feat: make slider scrollable
